Started Refactor of Calculator
