+++ calc/calculations/multiplication.py
@@ -1,0 +1,7 @@
+"""Multiplication Class"""
+
+class Multiplication:
+
+    @staticmethod
+    def multiply(value_a, value_b):
+        return value_a * value_b
