+++ calc/calculations/subtraction.py
@@ -1,6 +1,7 @@
 """Subtraction Class"""
 
 class Subtraction:
+    """Subtracting two numbers"""
 
     @staticmethod
     def subtract(value_a, value_b):
