+++ calc/calculator.py
@@ -1,16 +1,24 @@
 """ This is the increment function"""
-from calc.operations.addition import Addition
-from calc.operations.subtraction import Subtraction
+from calc.calculations.addition import Addition
+from calc.calculations.subtraction import Subtraction
+from calc.calculations.multiplication import Multiplication
 
 class Calculator:
     """ This is the Calculator class"""
+    history = []
 
     @staticmethod
     def add_numbers(value_a, value_b):
         """ adds two numbers"""
-        return Addition.add(value_a, value_b)
+        addition = Addition(value_a, value_b)
+        Calculator.history.append(addition)
+        return addition.getResult()
 
     @staticmethod
     def subtract_numbers(value_a, value_b):
         """ subtract number from result"""
         return Subtraction.subtract(value_a,value_b)
+    @staticmethod
+    def multiply_numbers(value_a, value_b):
+        """ subtract number from result"""
+        return Multiplication.multiply(value_a,value_b)
