--- calc/operations/addition.py
@@ -1,7 +1,0 @@
-"""Addition Class"""
-
-class Addition:
-
-    @staticmethod
-    def add(value_a, value_b):
-        return value_a + value_b
