+++ tests/addition_test.py
@@ -1,6 +1,2 @@
 """Testing Addition"""
-from calc.operations.addition import Addition
-
-def test_addition():
-    """testing calc result is 0"""
-    assert Addition.add(1,2) == 3
+from calc.calculations.addition import Addition
